Add tensorflow BatchNorm float16 support (#9582)
Adds support for float16 inputs to the BatchNormalization layer. To
support float16, BatchNormalization stores its weights as float32 which
allows for more precision when performing the weight updates. It then
scales the weights down to float16 during forward and backward
propagation. This follows and NVIDIA's recommendation and TensorFlow's
implementation.

+++ keras/engine/topology.py
@@ -410,7 +410,7 @@ class Layer(object):
         initializer = initializers.get(initializer)
         if dtype is None:
             dtype = K.floatx()
-        weight = K.variable(initializer(shape),
+        weight = K.variable(initializer(shape, dtype=dtype),
                             dtype=dtype,
                             name=name,
                             constraint=constraint)
